zsh: Fix issues with prompt async_fd reuse
Fix issues where the prompt async_fd is reused after we are done with
it. We check if the file descriptor is alive then try to remove the
handle and close it, even if we are no longer the one using the file
descriptor.

Fix the issues by unsetting the async_fd variable when we are done with
it, thus we no longer attempt to close the file descriptor when we are
not using it.