kitty: Add dropdown terminal wrapper for kitty
Use tdrop to create a wrapper shell program that uses kitty as if it
were a dropdown terminal.

This is combined with custom window behaviour and properties in KDE to
create the desired aesthetic, such as removing the titlebar but leaving
in window borders and custom colouring for thw window borders.